chore: refactor floating variables suggestions into multiple components
