+++ src/components/common/ClipboardButton/types.ts
@@ -1,0 +1,3 @@
+export default interface ClipboardProps {
+    content: string
+}
