+++ src/components/common/index.ts
@@ -37,3 +37,5 @@ export * from './DetectBottom'
 export * from './hooks/ClickOutside'
 export * from './hooks/FileReader'
 export { default as FloatingVariablesSuggestions } from './FloatingVariablesSuggestions/FloatingVariablesSuggestions'
+export { default as CopyClipboard } from './ClipboardButton/ClipboardButton'
+export { default as DebouncedSearch } from './DebouncedSearch/DebouncedSearch'
